CryptoTradingBot/requirements.txt
python-binance
TA-lib
numpy
websocket-client